initial commit with refactor and restructure to conform to new app structure
