+++ merino/providers/top_picks/__init__.py
@@ -1,0 +1,1 @@
+"""Top Picks Provider"""
