+++ merino/providers/top_picks/backends/top_picks.py
@@ -1,0 +1,1 @@
+"""A wrapper for Top Picks Provider I/O Interactions"""
